[#130320041] Make it easier to differentiate attributes in case table
Attributes with long names may have unique part at end.
Make case table header accommodate two lines of text.
Make header truncate text in the middle so that the beginning of the name and end are always visible.

This change was made 9/23/16 in the libraries/SlickGrid/ directory of
CODAP and not synched with this version. This commit corrects that
omission.